Added Json versions of adding ACF field groups
So I tried so see if I could create snippets of ACF fields in json to
have something I could import and export quickly. And something that
would work with the api. I think I'll stop working on this for now and
switch to another branch and make them in php. Then maybe in json.